requirements.txt
requests
beautifulsoup4
pandas
fastapi
uvicorn
python-dotenv
google-cloud-bigquery
# google-cloud-aiplatform
# google-genai
pyarrow
anthropic